New search into master.

+++ platform/http_request.cpp
@@ -1,6 +1,7 @@
-#include "platform/http_request.hpp"
 #include "platform/chunks_download_strategy.hpp"
+#include "platform/http_request.hpp"
 #include "platform/http_thread_callback.hpp"
+#include "platform/platform.hpp"
 
 #include "defines.hpp"
 
@@ -277,7 +278,8 @@ class FileHttpRequest : public HttpRequest, public IHttpThreadCallback
         (void)my::DeleteFileX(m_filePath + RESUME_FILE_EXTENSION);
 
         // Rename finished file to it's original name.
-        (void)my::DeleteFileX(m_filePath);
+        if (Platform::IsFileExistsByFullPath(m_filePath))
+          (void)my::DeleteFileX(m_filePath);
         CHECK(my::RenameFileX(m_filePath + DOWNLOADING_FILE_EXTENSION, m_filePath), ());
 
         DisableBackupForFile(m_filePath);
@@ -394,22 +396,6 @@ HttpRequest * HttpRequest::PostJson(string const & url, string const & postData,
   return new MemoryHttpRequest(url, postData, onFinish, onProgress);
 }
 
-namespace
-{
-  class ErrorHttpRequest : public HttpRequest
-  {
-    string m_filePath;
-  public:
-    ErrorHttpRequest(string const & filePath)
-      : HttpRequest(CallbackT(), CallbackT()), m_filePath(filePath)
-    {
-      m_status = EFailed;
-    }
-
-    virtual string const & Data() const { return m_filePath; }
-  };
-}
-
 HttpRequest * HttpRequest::GetFile(vector<string> const & urls,
                                    string const & filePath, int64_t fileSize,
                                    CallbackT const & onFinish, CallbackT const & onProgress,
@@ -423,13 +409,8 @@ HttpRequest * HttpRequest::GetFile(vector<string> const & urls,
   {
     // Can't create or open file for writing.
     LOG(LWARNING, ("Can't create file", filePath, "with size", fileSize, e.Msg()));
-
-    // Mark the end of download with error.
-    ErrorHttpRequest error(filePath);
-    onFinish(error);
-
-    return 0;
   }
+  return nullptr;
 }
 
 } // namespace downloader
